+++ skbio/tree/_tree.py
@@ -3269,3 +3269,407 @@ class TreeNode(SkbioObject):
                 nodes_to_unpack.append(node)
         for node in nodes_to_unpack:
             node.unpack()
+
+    @experimental(as_of="0.5.3")
+    def is_ultrametric(self):
+        """Returns whether the tree is ultrametric or not.
+
+        Returns
+        -------
+        bool
+            True is all root-to-tip distance are equal, or False if at least
+            one single distance differs
+
+        Notes
+        -----
+        This function collects all the tip to root branch lengths in a set.
+        If this set has a length of 1, then the tree is ultrametric.
+
+        Raises
+        ------
+        NoLengthError
+            A NoLengthError will be thrown if a node without length is
+            encountered
+
+        See Also
+        --------
+        distance
+        tip_tip_distances
+
+        Examples
+        --------
+        >>> from six import StringIO
+        >>> from skbio import TreeNode
+        >>> # example for a non-ultrametric tree
+        >>> tree = TreeNode.read(["((a:1,b:2)c:3,(d:4,e:5)f:6)root;"])
+        >>> tree.is_ultrametric()
+        False
+        >>> tip_root = tree.root()
+        >>> print([tip_root.distance(tree.find(tip)) for tip in tree.tips()])
+        [4.0, 5.0, 10.0, 11.0]
+        >>> # example for an ultrametric tree
+        >>> tree = TreeNode.read(['((a:3,b:3)e:7,(c:2,d:2)f:8);'])
+        >>> tree.is_ultrametric()
+        True
+        >>> tip_root = tree.root()
+        >>> print([tip_root.distance(tree.find(tip)) for tip in tree.tips()])
+        [10.0, 10.0, 10.0, 10.0]
+        """
+        tip_root = self.root()
+        metrics = set()
+        for tip in self.tips():
+            root_to_tip_dist = tip_root.distance(self.find(tip))
+            metrics.add(root_to_tip_dist)
+        if len(metrics) > 1:
+            return False
+        return True
+
+    @experimental(as_of="0.5.3")
+    def scale_length(self):
+        """Returns the tree where the branch lengths are
+        scaled, i.e. divided by (i) the root-to-tip distance
+        for ultrametric trees and (ii) the total branch
+        length (method `descending_branch_length`) for
+        non-ultrametric trees.
+
+        Returns
+        -------
+        TreeNode
+            The length-scaled tree
+
+        Raises
+        ------
+        NoLengthError
+            A NoLengthError will be thrown if a node without length is
+            encountered (from the `distance` method)
+
+        See Also
+        --------
+        distance
+        is_ultrametric
+
+        Examples
+        --------
+        >>> from six import StringIO
+        >>> from skbio import TreeNode
+        >>> # scaling non-ultrametric tree uses total branch length
+        >>> t = TreeNode.read(["((a:1,b:2)c:3,d:4)root;"])
+        >>> t2 = t.scale_length()
+        >>> [(n.length, t2.find(n.name).length) for n in t.preorder(False)]
+        [(3.0, 0.3), (1.0, 0.1), (2.0, 0.2), (4.0, 0.4)]
+        >>> # scaling ultrametric tree uses root-to-tip
+        >>> t = TreeNode.read(['((a:3,b:3)e:7,f:10)root;'])
+        >>> t2 = t.scale_length()
+        >>> [(n.length, t2.find(n.name).length) for n in t.preorder(False)]
+        [(7.0, 0.7), (3.0, 0.3), (3.0, 0.3), (10.0, 1.0)]
+        """
+        tcopy = self.deepcopy()
+        if tcopy.is_ultrametric():
+            # if tree is ultrametric, use longest edge
+            tip_root = tcopy.root()
+            for tip in tcopy.tips():
+                ref_branch_length = tip_root.distance(tcopy.find(tip))
+                break
+        else:
+            # if tree is not ultrametric, use total branch length
+            ref_branch_length = tcopy.descending_branch_length()
+        for node in tcopy.levelorder(include_self=False):
+            if node.length is None:
+                raise NoLengthError("No length on node %s found!" %
+                                    node.name or "unnamed")
+            node.length = node.length / ref_branch_length
+        return tcopy
+
+    @experimental(as_of="0.5.3")
+    def _get_nodes_degrees(self):
+        """Returns the out degree of each node.
+
+        Returns
+        -------
+        list
+            Number of nodes furcating out (degree) of each node along preorder.
+
+        Notes
+        -----
+        First collects the list of `[parent_node, current_node]` names in a
+        numpy array and crate a dict with count the number of times (values)
+        each parent occurs in the `parent_node` column of the array (keys).
+        Then, create the degree list for each node, with the dict for internal
+        nodes and 1 for tips.
+
+        Example
+        -------
+        >>> # preorder here is: a, c, d, b, e, f
+        >>> tree = TreeNode.read(['((c,d)a,(e,f)b);'])
+        >>> tree._get_nodes_degrees()
+        [2, 1, 1, 2, 1, 1]
+        """
+        parents_nodes = np.array(
+            [[n.parent.name, n.name] if n.parent.name else ['root', n.name]
+                for n in self.preorder(include_self=False)])
+        names, counts = np.unique(parents_nodes[:, 0], return_counts=True)
+        nparents = dict(zip(names, counts))  # degree of internal node
+        # make it a per node vector and set it to 1 for tips
+        degree = [nparents.get(x, 1) for x in parents_nodes[:, 1]]
+        return degree
+
+    @experimental(as_of="0.5.3")
+    def _get_preorder_array(self, tcopy):
+        """Create a numpy array version of the preorder tree with nodes as
+        rows and as columns: parent name, node name, branch length,
+        and number of children.
+
+        Parameters
+        ----------
+        tcopy : TreeNode or None
+            A copy of self with modified branch lengths (or not)
+
+        Returns
+        -------
+        numpy.array
+            for each node (incl. tips):
+                [0] name of the parent
+                [1] name of the node
+                [2] length of the parent->node branch
+                [3] number of children for the node (tips are given one child)
+
+        Notes
+        -----
+        The number of children for a tips is set to 1, because this is used
+        as denominator in the evolutionary distinctiveness calculation.
+
+        Raises
+        ------
+        NoLengthError
+            A NoLengthError will be thrown if a node without length is
+            encountered.
+
+        See Also
+        --------
+        scale_length
+        evol_distinct
+        """
+        tab = []
+        for n in self.preorder(include_self=False):
+
+            if n.length is None:
+                raise NoLengthError("No length on node %s found." %
+                                    n.name or "unnamed")
+
+            tab.append([
+                n.parent.name,
+                n.name,
+                tcopy.find(n.name).length,
+                ((1 * n.is_tip()) + n.count(tips=True))
+            ])
+        return np.array(tab)
+
+    @experimental(as_of="0.5.3")
+    def name_unnamed_internal_nodes(self):
+        """Name the internal nodes that do not have a name.
+
+        Notes
+        -----
+        First, the existing node names are collected in a set, and then the
+        tree is traversed in preorder so that the unnamed nodes are given an
+        incremental `N#` name.
+
+        See Also
+        --------
+        evol_distinct
+
+        Example
+        -------
+        >>> tree = TreeNode.read(['((c,d),(e));'])
+        >>> tree.name_unnamed_internal_nodes()
+        >>> str(tree).strip()
+        '((c,d)N2,(e)N3)N1;'
+        >>> tree = TreeNode.read(['((c,d)N1,(e));'])
+        >>> tree.name_unnamed_internal_nodes()
+        >>> str(tree).strip()
+        '((c,d)N1,(e)N3)N2;'
+        """
+        offset = 0
+        nodes_names = set([x.name for x in self.preorder() if x.name])
+        for curr in self.preorder():
+            if curr.is_tip() or curr.name:
+                continue
+            offset += 1
+            name = 'N%s' % offset
+            while name in nodes_names:
+                offset += 1
+                name = 'N%s' % offset
+            curr.name = name
+
+    @experimental(as_of="0.5.3")
+    def _calc_equal_splits(self, tab):
+        """Returns the evolutionary distinctiveness per node for the case
+        where the node furcations on the path to a tip are weighted in terms
+        of number of direct children at that node, i.e. will always be `2`
+        for each node on a strictly bifurcating tree.
+
+        Parameters
+        ----------
+        tab : numpy.array
+            each node's parent name, name, branch length and number of children
+
+        Returns
+        -------
+        dict
+            Evolutionary distinctiveness (values) of nodes (keys)
+
+        Notes
+        -----
+        Collects the degree of each node (tips as having one child since
+        dividing per one for the ED calculation is needed to not have to
+        divide by zero), and increments a per-node zero-initialized list
+        with, along the preorder:
+          `(parent_node_ED + node_branch_length / number of children)`
+        Python re-coding of the R package `phyloregion::evol_distinct`
+        https://rdrr.io/cran/phyloregion/src/R/evol_distinct.R
+
+        See Also
+        --------
+        evol_distinct
+        """
+        root_name = self.root().name
+        degree = self._get_nodes_degrees()
+        # evolutionary distinctiveness (ED) result vector (to be populated)
+        res = [0] * tab.shape[0]
+        # per each node along the tabulated preorder (cladewise)
+        for idx_node, row in enumerate(tab):
+            # update the ED result by adding the branch length
+            # and dividing the updated value by the descending
+            # degree (always `2` for a strictly bifurcating tree)
+            if row[0] == root_name:
+                base = 0  # because the root in not present in the ED vector
+            else:
+                base = res[list(tab[:, 1]).index(row[0])]
+            res[idx_node] = (float(base) + float(row[2])) / degree[idx_node]
+        return dict(zip(tab[:, 1], res))
+
+    @experimental(as_of="0.5.3")
+    def _calc_fair_proportion(self, tab):
+        """Returns the evolutionary distinctiveness per node for the case
+        where the node furcations on the path to a tip are weighted in terms
+        of total number of tips down that node.
+
+        Parameters
+        ----------
+        tab : numpy.array
+            each node's parent name, name, branch length and number of children
+
+        Returns
+        -------
+        dict
+            Evolutionary distinctiveness (values) of nodes (keys)
+
+        Notes
+        -----
+        Collects the degree of each node (tips as having one child since
+        dividing per one for the ED calculation is needed to not have to
+        divide by zero), and increments a per-node zero-initialized list
+        with, along the preorder:
+          `parent_node_ED + (node_branch_length / total number of tips)`
+        Python re-coding of the R package `phyloregion::evol_distinct`
+        https://rdrr.io/cran/phyloregion/src/R/evol_distinct.R
+
+        See Also
+        --------
+        evol_distinct
+        """
+        root_name = self.root().name
+        # evolutionary distinctiveness (ED) result vector (to be populated)
+        res = [0] * tab.shape[0]
+        # per each node along the tabulated preorder (cladewise)
+        for idx_node, row in enumerate(tab):
+            # divide the node's branch length by its total number of tips and
+            # add this to update the ED result
+            if row[0] == root_name:
+                base = 0  # because the root in not present in the ED vector
+            else:
+                base = res[list(tab[:, 1]).index(row[0])]
+            res[idx_node] = float(base) + (float(row[2]) / float(row[3]))
+        return dict(zip(tab[:, 1], res))
+
+    @experimental(as_of="0.5.3")
+    def evol_distinct(self, algorithm="fair_proportion",
+                      scale=False, use_branch_lengths=True):
+        """Returns the evolutionary distinctiveness (ED) of the tips of a tree.
+
+        Parameters
+        ----------
+        algorithm : str
+            Either of "fair_proportion" or "equal_splits", to control how the
+            nodes furcations are used
+        scale : bool
+            Whether to normalize branch length before ED calculation
+            (false by default)
+        use_branch_lengths : bool
+            Whether to use branch length for ED calculation (true by default)
+
+        Returns
+        -------
+        dict
+            Evolutionary distinctiveness (values) of tips (keys)
+
+        Raises
+        ------
+        NoLengthError
+            A NoLengthError will be thrown if a node without length is
+            encountered.
+        ValueError
+            A ValueError will be thrown if the algorithm name is neither
+            "equal_splits" or "fair_proportion".
+
+        Notes
+        -----
+        Python re-coding of the R package `phyloregion::evol_distinct`
+        https://rdrr.io/cran/phyloregion/src/R/evol_distinct.R
+
+        See Also
+        --------
+        name_unnamed_internal_nodes
+        calc_equal_splits
+        calc_fair_proportion
+        scale_length
+
+        Examples
+        --------
+        >>> from six import StringIO
+        >>> from skbio import TreeNode
+        >>> tree = TreeNode.read(["(((a:9,b:12)f:16,c:4)e:72,d:9)root;"])
+        >>> print(tree.ascii_art())
+                                      /-a
+                            /f-------|
+                  /e-------|          \\-b
+                 |         |
+        -root----|          \\-c
+                 |
+                  \\-d
+        >>> # the ED of a tip furcating from the root with be its branch length
+        >>> tree.evol_distinct(algorithm='equal_splits')
+        {'a': 35.0, 'b': 38.0, 'c': 40.0, 'd': 9.0}
+        >>> tree.evol_distinct(algorithm='fair_proportion')
+        {'a': 41.0, 'b': 44.0, 'c': 28.0, 'd': 9.0}
+        """
+        self.name_unnamed_internal_nodes()  # make sure all nodes have a name
+        # pre-treatment based on parameters
+        if scale:
+            tcopy = self.scale_length()  # branch length normalization
+        else:
+            tcopy = self.deepcopy()  # or just make a copy
+        # branch length homogenization (rely on furcation info only)
+        if not use_branch_lengths:
+            for node in tcopy.levelorder(include_self=False):
+                node.length = 1
+        # ED calculation involves laying the preorder tree in a numpy array
+        tab = self._get_preorder_array(tcopy)
+        if algorithm == "equal_splits":
+            res = self._calc_equal_splits(tab)  # use out degree of nodes
+        elif algorithm == "fair_proportion":
+            res = self._calc_fair_proportion(tab)  # use tip numbers of nodes
+        else:
+            raise ValueError('Invalid algorithm name: "%s"' % algorithm)
+        res_d = dict((x.name, round(res[x.name], 8)) for x in tcopy.tips())
+        return res_d