+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Features
 
+* Added `evol_distinct` method to measure Evolutionary Distinctiveness (ED), [issue #1777](https://github.com/biocore/scikit-bio/issues/1777):
+  * private methods `_calc_fair_proportion` and `_calc_equal_splits` to calculate ED under two different node furcation modes
+  * private method `_get_preorder_array` to make the table use under the hood for ED calculation
+  * private method `_get_nodes_degrees` to get the out degree per internal node
+  * method `name_unnamed_internal_nodes` to name unnamed nodes (necessary for ED calculation)
+  * method `is_ultrametric` to check is tree is ultrametric or not
+  * method `scale_length` to normalize branch length
 * Added support for `np.float32` with `DissimilarityMatrix` objects.
 * Added support for method and number_of_dimensions to permdisp reducing the runtime by 100x at 4000 samples, [issue #1769](https://github.com/biocore/scikit-bio/pull/1769).
 * OrdinationResults object is now accepted as input for permdisp.